requirements.txt
# Vercel API requirements
Pillow>=10.0.0
